made table reuseable
